requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
django-jazzmin==2.6.0
django-jet==1.0.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
Pillow==10.0.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
